% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsampleReach.R
\name{subsampleReach}
\alias{subsampleReach}
\title{Subsample Reach}
\usage{
subsampleReach(rch_sf = NULL, start_pt = NULL, needed_length = 0.5)
}
\arguments{
\item{rch_sf}{\code{sf} object containing the line layer of a possible GRTS reach.}

\item{start_pt}{\code{sf} object containing the starting point (e.g. GRTS point) where reach will be built around}

\item{needed_length}{distance (in miles) that the reach should extend on one side of the GRTS point.}
}
\value{
sf
}
\description{
Starting from a fixed point, extracts a subsample of a reach in one direction from that point, extending it for a specified distance.
}
\examples{
subsampleReach()
}
\author{
Kevin See
}
